requirements.txt
PyPDF2==3.0.1
langchain-text-splitters==0.3.4
fastapi==0.115.6
openai==1.58.1
pinecone==5.4.2
python-dotenv==1.0.1
pydantic==2.10.4
langchain-community==0.3.13
uvicorn==0.34.0
tiktoken==0.8.0
fastapi-cors==0.0.6
python_multipart==0.0.20